+++ apps/web/uno.config.ts
@@ -1,0 +1,13 @@
+import { defineConfig, presetIcons } from "unocss";
+
+export default defineConfig({
+  presets: [
+    presetIcons({
+      extraProperties: {
+        display: "inline-block",
+        "vertical-align": "middle",
+      },
+      warn: true,
+    }),
+  ],
+});